+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Added `pybricksdev.ble.pybricks.Command.PBIO_PYBRICKS_COMMAND_REBOOT_TO_UPDATE_MODE`.
+
+### Fixed
+- Fixed reboot in update mode for Pybricks Profile >= 1.2.0 in `pybricksdev flash` CLI.
+
 ## [1.0.0-alpha.33] - 2022-11-06
 
 ### Changed